Move setter attribute in PropertyDescriptor proper
