+++ spring-boot-project/spring-boot-tools/spring-boot-configuration-processor/src/main/java/org/springframework/boot/configurationprocessor/JavaBeanPropertyDescriptor.java
@@ -21,8 +21,6 @@ import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.VariableElement;
 import javax.lang.model.type.TypeMirror;
 
-import org.springframework.boot.configurationprocessor.metadata.ItemDeprecation;
-
 /**
  * A {@link PropertyDescriptor} for a standard JavaBean property.
  *
@@ -30,17 +28,10 @@ import org.springframework.boot.configurationprocessor.metadata.ItemDeprecation;
  */
 class JavaBeanPropertyDescriptor extends PropertyDescriptor<ExecutableElement> {
 
-	private final ExecutableElement setter;
-
 	JavaBeanPropertyDescriptor(TypeElement ownerElement, ExecutableElement factoryMethod,
 			ExecutableElement getter, String name, TypeMirror type, VariableElement field,
 			ExecutableElement setter) {
-		super(ownerElement, factoryMethod, getter, name, type, field, getter);
-		this.setter = setter;
-	}
-
-	public ExecutableElement getSetter() {
-		return this.setter;
+		super(ownerElement, factoryMethod, getter, name, type, field, getter, setter);
 	}
 
 	@Override
@@ -48,15 +39,6 @@ class JavaBeanPropertyDescriptor extends PropertyDescriptor<ExecutableElement> {
 		boolean isCollection = env.getTypeUtils().isCollectionOrMap(getType());
 		return !env.isExcluded(getType()) && getGetter() != null
 				&& (getSetter() != null || isCollection);
-	}
-
-	@Override
-	protected ItemDeprecation resolveItemDeprecation(
-			MetadataGenerationEnvironment environment) {
-		boolean deprecated = environment.isDeprecated(getGetter())
-				|| environment.isDeprecated(getSetter())
-				|| environment.isDeprecated(getFactoryMethod());
-		return deprecated ? environment.resolveItemDeprecation(getGetter()) : null;
 	}
 
 }
