+++ spring-boot-project/spring-boot-tools/spring-boot-configuration-processor/src/main/java/org/springframework/boot/configurationprocessor/PropertyDescriptorResolver.java
@@ -61,8 +61,9 @@ class PropertyDescriptorResolver {
 		members.getFields().forEach((name, field) -> {
 			TypeMirror returnType = field.asType();
 			ExecutableElement getter = members.getPublicGetter(name, returnType);
+			ExecutableElement setter = members.getPublicSetter(name, returnType);
 			candidates.add(new LombokPropertyDescriptor(type, factoryMethod, field, name,
-					returnType, getter));
+					returnType, getter, setter));
 		});
 		return candidates.stream().filter(this::isCandidate);
 	}
